feat: temper my expectations / only abstract the incoming data stream handling code, and matching other sdk implementations
